Add a script that makes a k8s deployment of a km snapshot
k8s_snapshot_workflow.sh is a script that will take a deployment/pod
yaml, update it so that a snapshot of the application can be taken.
Verify that the application is running.
Take the snapshot.
Delete the pod/deployment.
Build a new container from the original container but run the snapshot.
Then make another deployment/pod definition that runs the snapshot.
Start that pod/deployment and verify that it works.
Delete the pod/deployment.

Also included is a sample yaml file I used for testing.

I doubt we want to really check this in.  It is being pushed to give
people a chance to look at it and decide how to make it fit our needs
better.